Rewrite the update client and updater.
