+++ partials/no-products.php
@@ -1,0 +1,3 @@
+<div class="no-products">
+	<p><?php esc_attr_e( 'No products in this category yet.', 'slswcclient' ); ?></p>
+</div>
